次元切り替えスイッチによってシーンのオブジェクトの追加削除を行うよう変更
