+++ app/index.html
@@ -12,10 +12,10 @@
     <div class="app">
       <div id="canvas"></div>
       <div class="settings">
-        <div class="dimension_toggle">
+        <label class="dimension_toggle">
           <input type="checkbox" id="dimension_toggle_switch" />
-          <div class="slider">2D/3D</div>
-        </div>
+          <div id="dimension_toggle_slider">3D</div>
+        </label>
         <div class="enable_auto_rotate">
           <input type="checkbox" id="checkbox_auto_rotate" checked />
           <label for="checkbox_auto_rotate">自動回転</label>
