+++ app/script/main.js
@@ -1,6 +1,7 @@
 import * as THREE from "three";
 import { TransformControls } from "three/addons/controls/TransformControls";
 import PointCharge from "./PointCharge.js";
+import { Field3D } from "./Field3D.js";
 
 import {
   CreateScene,
@@ -14,43 +15,12 @@ const renderer = CreateRenderer(document.getElementById("canvas"));
 const camera = CreateCamera();
 const controls = CreateControls(camera, renderer);
 
-// document
-//   .getElementById("dimension_toggle_switch")
-//   .addEventListener("change", (e) => {
-//     if (e.target.checked) {
-//       camera.position.set(0, 0, 100);
-//       camera.lookAt(0, 0, 0);
-//     } else {
-//       camera.position.set(100, 100, 100);
-//       camera.lookAt(0, 0, 0);
-//     }
-//   });
-
 // 自動回転の有無チェックイベント
 document
   .getElementById("checkbox_auto_rotate")
   .addEventListener("change", (e) => {
     controls.autoRotate = e.target.checked;
   });
-
-// 指定座標における電場ベクトルを計算
-const PosToElectricFieldVector = (pos, point_charges) => {
-  var electric_field_vector = new THREE.Vector3(0, 0, 0); // 合成ベクトル
-  for (const point_charge of point_charges) {
-    const diff = new THREE.Vector3(0, 0, 0);
-    diff.subVectors(pos, point_charge.pos); // 点電荷と観測点との差分ベクトル
-
-    const r_sq = diff.lengthSq(); // 点電荷と観測点との距離^2 (平方根処理をなくすため)
-    const k = 8.987552 * 10 ** 9; // クーロン定数
-
-    const factor = r_sq ? (k * point_charge.charge) / r_sq ** 2 : 0; // ベクトルにかける係数
-
-    diff.multiplyScalar(factor);
-
-    electric_field_vector.add(diff);
-  }
-  return electric_field_vector;
-};
 
 // オブジェクトコントローラー
 const transControls = new TransformControls(camera, renderer.domElement);
@@ -63,6 +33,24 @@ const point_charges = [
   new PointCharge(+0.00001, new THREE.Vector3(-50, 0, 20)),
   new PointCharge(0.00005, new THREE.Vector3(20, 0, 0)),
 ];
+
+const field_3d = new Field3D(point_charges);
+scene.add(field_3d);
+
+// 2D/3D切り替え
+document
+  .getElementById("dimension_toggle_switch")
+  .addEventListener("change", (e) => {
+    if (e.target.checked) {
+      // 2D
+      document.getElementById("dimension_toggle_slider").textContent = "2D";
+      scene.remove(field_3d);
+    } else {
+      // 3D
+      document.getElementById("dimension_toggle_slider").textContent = "3D";
+      scene.add(field_3d);
+    }
+  });
 
 // 球
 const sphereGeometry = new THREE.SphereGeometry(1, 32, 32);
@@ -84,59 +72,6 @@ for (const charge of point_charges) {
   // transControls.attach(mesh);
 }
 
-/// @param origin 始点 (PointCharge)
-const CreateElectricLineGeometry = (
-  origin_charge,
-  vector,
-  point_charges,
-  length
-) => {
-  const points = [origin_charge.pos.clone()];
-  const _origin = origin_charge.pos.clone();
-  _origin.add(vector);
-  for (var i = 0; i < length; i++) {
-    const d_vector = PosToElectricFieldVector(_origin, point_charges);
-    d_vector.normalize();
-    if (origin_charge.charge < 0) d_vector.multiplyScalar(-1);
-    // 点電荷との衝突判定
-    for (const point_charge of point_charges) {
-      const diff = new THREE.Vector3(0, 0, 0);
-      diff.subVectors(_origin, point_charge.pos);
-      if (diff.length() < 1) {
-        console.log("hit");
-        return new THREE.BufferGeometry().setFromPoints(points);
-      }
-    }
-    _origin.add(d_vector);
-    points.push(_origin.clone());
-  }
-  return new THREE.BufferGeometry().setFromPoints(points);
-};
-
-const line_material = new THREE.LineBasicMaterial({ color: 0x0000ff });
-for (const point_charge of point_charges) {
-  for (var n_theta = 0; n_theta < 10; n_theta++) {
-    for (var n_phi = 0; n_phi < 10; n_phi++) {
-      const theta = ((Math.PI * 2) / 10) * n_theta;
-      const phi = ((Math.PI * 2) / 10) * n_phi;
-      const x = 5 * Math.sin(theta) * Math.cos(phi);
-      const y = 5 * Math.sin(theta) * Math.sin(phi);
-      const z = 5 * Math.cos(theta);
-
-      const line = new THREE.Line(
-        CreateElectricLineGeometry(
-          point_charge,
-          new THREE.Vector3(x, y, z),
-          point_charges,
-          1000
-        ),
-        line_material
-      );
-      scene.add(line);
-    }
-  }
-}
-
 const CreateElectricFieldVector = () => {
   var arrows = [];
   for (var point_charge of point_charges) {
@@ -147,7 +82,7 @@ const CreateElectricFieldVector = () => {
           if (x === 0 && y === 0 && z === 0) continue;
           const origin = new THREE.Vector3(x * 5, y * 5, z * 5);
           origin.add(point_charge.pos);
-          const electric_field_vector = PosToElectricFieldVector(
+          const electric_field_vector = field_3d.posToElectricFieldVector(
             origin,
             point_charges
           );
